pyproject.toml
[project]
name = "contractfol"
version = "0.1.0"
description = "Validação Automatizada de Contratos Inter-Institucionais usando LLMs e Lógica de Primeira Ordem"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Anderson Rezende", email = "anderson@example.com"}
]
keywords = [
    "contract-validation",
    "llm",
    "first-order-logic",
    "z3",
    "neurosymbolic-ai",
    "legal-nlp"
]

dependencies = [
    "z3-solver>=4.12.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "spacy>=3.7.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "pydantic>=2.5.0",
    "python-docx>=1.1.0",
    "pypdf2>=3.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.4.0",
    "datasets>=2.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
contractfol = "contractfol.cli:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
